+++ src/dcqc/tests/__init__.py
@@ -1,0 +1,10 @@
+from dcqc.tests.base import BaseTest, ExternalTestMixin, Process, TestStatus
+from dcqc.tests.bioformats_info_test import BioFormatsInfoTest
+from dcqc.tests.file_extension_test import FileExtensionTest
+from dcqc.tests.grep_date_test import GrepDateTest
+from dcqc.tests.json_load_test import JsonLoadTest
+from dcqc.tests.jsonld_load_test import JsonLdLoadTest
+from dcqc.tests.libtiff_info_test import LibTiffInfoTest
+from dcqc.tests.md5_checksum_test import Md5ChecksumTest
+from dcqc.tests.ome_xml_schema_test import OmeXmlSchemaTest
+from dcqc.tests.tiff_tag_306_date_time_test import TiffTag306DateTimeTest
